estudiolapa/requirements.txt
Django
psycopg2
ipython
ipdb
Pillow
python-thumbnails
gunicorn
django-bootstrap4
